Added pricing sheet inch foot form
